Removing un needed changes for program enrollment (#249)
