+++ api/src/main/resources/messages.properties
@@ -61,13 +61,3 @@ ugandaemr.encounter.changedby=Changed By
 ugandaemr.encounter.datechanged=Changed On
 
 coreapps.findPatient.registerPatient.label=Register Patient
-
-aijar.app.program.title=Program
-aijar.app.program.selectPatient=Select Patient
-aijar.app.program.label=Program
-aijar.program.enroll=Enroll
-aijar.program.enrollmentDate=Enrollment Date
-aijar.program.entryPoint=Entry Point
-aijar.program.completionDate=Completion Date
-aijar.program.outcome=Outcome
-aijar.program.title=Patient Program
